+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Documentaion and examples on [tarantool/http server](https://github.com/tarantool/http) middleware
 
-###Fixed
+### Fixed
 - Throw exception when `http_middleware.build_default_collector` is called with same name
 
 ## [0.2.0] - 2020-05-07